code reorganisation and documentation
